pyproject.toml
[tool.poetry]
name = "bewise-ai-questions-api"
version = "0.1.0"
description = ""
authors = ["Timur <81976283+mym1chelle@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.95.2"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.15"}
alembic = "^1.11.1"
asyncpg = "^0.27.0"
aiohttp = "^3.8.4"
python-dateutil = "^2.8.2"
environs = "^9.5.0"
gunicorn = "^20.1.0"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.2"
flake8 = "^6.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
